Add GUI HomepageView without clusters inside
Here I still don't know how to handle the display of clusters, so they are not
added. It is problematic to add them from the top left corner going to the
right corner occasionally going to a new line. It is problematic to add such a
complicated element. I haven't figured out how to accomplish this yet.