Add ViewStack implementation - close #24
Adds an implementation of ViewStack that suffers from a couple of problems at
this point. The window at this point is not resizable until something is
invented that will nicely resize the objects in the QStackedWidget.

I have commented out the unnecessary classes for the moment, the current code
serves as an overview for further work.