Structure the project for test writing - close #29
An earlier version of the project was not set up for writing tests. From now it
is easy to create them. We were forced to use QTest to make it fairly easy to
test the GUI and write UITests. For this reason, we use QTests, but we are a
bit saddened by their poorness. Inability to write mocks or more complicated
tests forced us to add gtest and gmock. The google libraries are pulled
directly from their repository, so they are not stored inside our repository.
NizerTests executes tests for both frameworks.